chore(provider-utils): move ToolResultContent to provider-utils
