+++ packages/anthropic/src/anthropic-tools.ts
@@ -1,25 +1,12 @@
+import { ToolResultContent } from '@ai-sdk/provider-utils';
 import { z } from 'zod';
 
-// Copied from ai package
 type ExecuteFunction<PARAMETERS, RESULT> =
   | undefined
   | ((
       args: PARAMETERS,
       options: { abortSignal?: AbortSignal },
     ) => Promise<RESULT>);
-
-// Copied from ai package
-export type ToolResultContent = Array<
-  | {
-      type: 'text';
-      text: string;
-    }
-  | {
-      type: 'image';
-      data: string; // base64 encoded png image, e.g. screenshot
-      mediaType?: string; // e.g. 'image/png';
-    }
->;
 
 const Bash20241022Parameters = z.object({
   command: z.string(),
