+++ packages/provider-utils/src/types/tool-result.ts
@@ -23,3 +23,20 @@ Result of the tool call. This is the result of the tool's execution.
      */
   result: RESULT;
 }
+
+export type ToolResultContent = Array<
+  | {
+      type: 'text';
+      text: string;
+    }
+  | {
+      type: 'image';
+      data: string; // base64 encoded png image, e.g. screenshot
+      mediaType?: string; // e.g. 'image/png';
+
+      /**
+       * @deprecated Use `mediaType` instead.
+       */
+      mimeType?: string; // e.g. 'image/png';
+    }
+>;
